pyproject.toml
[project]
name = "prepzen-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "python-jose==3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
